requirements.txt
pynput>=1.7.7
pyobjc-framework-Cocoa
pyobjc-framework-Quartz

